booster reward checking improvement

+++ src/constants/models.ts
@@ -1,9 +1,9 @@
 import { Semblance } from '@semblance/structures';
 import { BoosterRewards, Information } from '@semblance/models';
-import { Message, TextChannel } from 'discord.js';
+import { GuildMember, Message, TextChannel } from 'discord.js';
 import config from '@semblance/config';
 import { formattedDate } from '.';
-const { sirhId, adityaId } = config;
+const { sirhId, adityaId, c2sGuildId } = config;
 
 // BoosterRewards - check dates for booster rewards
 export const checkBoosterRewards = async (client: Semblance) => {
@@ -23,20 +23,29 @@ export const checkBoosterRewards = async (client: Semblance) => {
     const promises = [];
     boosterRewards.forEach(async boosterReward => {
         const ogCodeLength = darwiniumCodes.list.length, darwiniumCode = darwiniumCodes.list.shift();
-        const user = await client.users.fetch(boosterReward.userId);
-        user.send(`Here's your code, ${user.username}! Code: ${darwiniumCode}`)
-        .catch(async err => {
-            await boosterChannel(client).send(`<@${user.id}> I had trouble DMing you so instead Aditya or SirH will manually provide you a code. :)`+
-            `\nTip: These errors tend to happen when your DMs are closed. So keeping them open would help us out :D`);
-            darwiniumCodes.list.unshift(darwiniumCode);
-        });
-        if (darwiniumCodes.list.length != ogCodeLength) promises.push(darwiniumCodes.update({ list: darwiniumCodes.list }));
-        promises.push(boosterReward.delete());
+        let failedToFetch: boolean, member: GuildMember;
+        try {
+            member = await client.guilds.cache.get(c2sGuildId).members.fetch(boosterReward.userId);
+        } catch (e) {
+            failedToFetch = true;
+        } finally {
+            if (failedToFetch) return promises.push(boosterReward.remove());        
+            if (!member.roles.cache.has(boosterRole(client).id)) return promises.push(boosterReward.remove());
+            member.user.send(`Here's your code, ${member.user.username}! Code: ${darwiniumCode}`)
+            .catch(async err => {
+                await boosterChannel(client).send(`${member} I had trouble DMing you so instead Aditya or SirH will manually provide you a code. :)`+
+                `\nTip: These errors tend to happen when your DMs are closed. So keeping them open would help us out :D`);
+                darwiniumCodes.list.unshift(darwiniumCode);
+            });
+            if (darwiniumCodes.list.length != ogCodeLength) promises.push(darwiniumCodes.update({ list: darwiniumCodes.list }));
+            promises.push(boosterReward.delete());
+        }
     });
     return Promise.all(promises);
 };
 
 export const boosterChannel = (client: Semblance) => client.channels.cache.get('547455179302109186') as TextChannel;
+export const boosterRole = (client: Semblance) => client.guilds.cache.get(c2sGuildId).roles.cache.get('660930089990488099');
 
 // BoosterRewards - create automatic booster rewards for author of message
 export const createBoosterRewards = async (message: Message) => {
